Add result formatter class
