+++ libstuff/SQResult.cpp
@@ -69,243 +69,6 @@ const string& SQResultRow::operator[](const string& key) const {
     STHROW_STACK("No column named " + key);
 }
 
-string SQResult::serializeToJSON() const {
-    // Just output as a simple object
-    // **NOTE: This probably isn't super fast, but could be easily optimized
-    //         if it ever became necessary.
-    STable output;
-    output["headers"] = SComposeJSONArray(headers);
-    vector<string> jsonRows;
-    for (size_t c = 0; c < rows.size(); ++c)
-        jsonRows.push_back(SComposeJSONArray(rows[c]));
-    output["rows"] = SComposeJSONArray(jsonRows);
-    return SComposeJSONObject(output);
-}
-
-string SQResult::serializeToText() const {
-    // Match the native format of sqlite3.
-    vector<size_t> maxLengths(headers.size());
-    for (size_t i = 0; i < headers.size(); i++) {
-        maxLengths[i] = headers[i].size();
-    }
-    for (size_t i = 0; i < rows.size(); i++) {
-        for (size_t j = 0; j < rows[i].size(); j++) {
-            if (rows[i][j].length() > maxLengths[j]) {
-                maxLengths[j] = rows[i][j].length();
-            }
-        }
-    }
-
-    // Create the output string
-    string output;
-
-    // Append the headers.
-    for (size_t i = 0; i < headers.size(); i++) {
-        string entry = headers[i];
-        entry.resize(maxLengths[i], ' ');
-        if (i != 0) {
-            output += "  ";
-        }
-        output += entry;
-    }
-    output += "\n";
-
-    // Now create the separator line.
-    for (size_t i = 0; i < maxLengths.size(); i++) {
-        string sep;
-        sep.resize(maxLengths[i], '-');
-        if (i != 0) {
-            output += "  ";
-        }
-        output += sep;
-    }
-    output += "\n";
-
-    // Finally, do each row.
-    for (size_t i = 0; i < rows.size(); i++) {
-        for (size_t j = 0; j < rows[i].size(); j++) {
-            string entry = rows[i][j];
-            entry.resize(maxLengths[j], ' ');
-            if (j != 0) {
-                output += "  ";
-            }
-            output += entry;
-        }
-        output += "\n";
-    }
-
-    // Done.
-    return output;
-}
-
-string SQResult::serializeTextDelimited(char delimiter) const {
-    auto needsQuoting = [&](const string& s) -> bool {
-        if (s.find('"') != string::npos) return true;
-        if (s.find('\n') != string::npos) return true;
-        if (s.find('\r') != string::npos) return true;
-        if (s.find(delimiter) != string::npos) return true;
-        return false;
-    };
-
-    auto quoteField = [&](const string& s) -> string {
-        if (!needsQuoting(s)) return s;
-        string out;
-        out.reserve(s.size() + 2);
-        out.push_back('"');
-        for (char c : s) {
-            if (c == '"') {
-                // CSV-style escape: double the quote
-                out.push_back('"');
-                out.push_back('"');
-            } else {
-                out.push_back(c);
-            }
-        }
-        out.push_back('"');
-        return out;
-    };
-
-    string output;
-
-    // Write headers if present
-    if (!headers.empty()) {
-        for (size_t i = 0; i < headers.size(); ++i) {
-            if (i) output.push_back(delimiter);
-            output += quoteField(headers[i]);
-        }
-        output.push_back('\n');
-    }
-
-    // Write rows
-    for (const auto& row : rows) {
-        size_t cols = headers.empty() ? row.size() : headers.size();
-        for (size_t j = 0; j < cols; ++j) {
-            if (j) output.push_back(delimiter);
-            string field = (j < row.size()) ? row[j] : string();
-            output += quoteField(field);
-        }
-        output.push_back('\n');
-    }
-
-    return output;
-}
-
-string SQResult::serializeToQuote() const {
-    auto isNumeric = [](const string& s) -> bool {
-        if (s.empty()) return false;
-        size_t i = 0;
-        // optional sign
-        if (s[i] == '+' || s[i] == '-') {
-            if (++i >= s.size()) return false;
-        }
-        bool anyDigits = false;
-        // integer part
-        while (i < s.size() && isdigit(static_cast<unsigned char>(s[i]))) {
-            anyDigits = true;
-            ++i;
-        }
-        // decimal part
-        if (i < s.size() && s[i] == '.') {
-            ++i;
-            while (i < s.size() && isdigit(static_cast<unsigned char>(s[i]))) {
-                anyDigits = true;
-                ++i;
-            }
-        }
-        if (!anyDigits) return false;
-        // exponent part
-        if (i < s.size() && (s[i] == 'e' || s[i] == 'E')) {
-            ++i;
-            if (i < s.size() && (s[i] == '+' || s[i] == '-')) ++i;
-            bool expDigits = false;
-            while (i < s.size() && isdigit(static_cast<unsigned char>(s[i]))) {
-                expDigits = true;
-                ++i;
-            }
-            if (!expDigits) return false;
-        }
-        return i == s.size();
-    };
-
-    auto quoteSQL = [](const string& s) -> string {
-        string out;
-        out.reserve(s.size() + 2);
-        out.push_back('\'');
-        for (char c : s) {
-            if (c == '\'') {
-                // SQL escape by doubling the quote
-                out.push_back('\'');
-                out.push_back('\'');
-            } else {
-                out.push_back(c);
-            }
-        }
-        out.push_back('\'');
-        return out;
-    };
-
-    const char delimiter = ','; // sqlite default separator (respects our CSV/TSV pattern)
-    string output;
-
-    // Emit headers if present (consistent with our CSV/TSV behavior)
-    if (!headers.empty()) {
-        for (size_t i = 0; i < headers.size(); ++i) {
-            if (i) output.push_back(delimiter);
-            output += quoteSQL(headers[i]);
-        }
-        output.push_back('\n');
-    }
-
-    for (const auto& row : rows) {
-        size_t cols = headers.empty() ? row.size() : headers.size();
-        for (size_t j = 0; j < cols; ++j) {
-            if (j) output.push_back(delimiter);
-            string field = (j < row.size()) ? row[j] : string();
-
-            // Literal NULL (unquoted) if exactly "NULL"
-            if (field == "NULL") {
-                output += "NULL";
-                continue;
-            }
-
-            // Numbers are bare (ASCII text, no quotes)
-            if (isNumeric(field)) {
-                output += field;
-                continue;
-            }
-
-            // Everything else is a SQL string literal
-            output += quoteSQL(field);
-        }
-        output.push_back('\n');
-    }
-
-    return output;
-}
-
-string SQResult::serializeToCSV() const {
-    return serializeTextDelimited(',');
-}
-
-string SQResult::serializeToTSV() const {
-    return serializeTextDelimited('\t');
-}
-
-string SQResult::serialize(SQResult::FORMAT format) const {
-    switch (format) {
-        case FORMAT::SQLITE3:
-            return serializeToText();
-        case FORMAT::CSV:
-            return serializeToCSV();
-        case FORMAT::TSV:
-            return serializeToTSV();
-        case FORMAT::JSON:
-            return serializeToJSON();
-        case FORMAT::QUOTE:
-            return serializeToQuote();
-    }
-}
-
 bool SQResult::deserialize(const string& json) {
     // Reset ourselves to start
     clear();
